refactor: changed files naming + removed unnecessary ingestion requests schemas
