+++ src/schemas/core/index.ts
@@ -3,3 +3,4 @@ export * from './job.schema';
 export * from './task.schema';
 export * from './mime.schema';
 export * from './aggregation.schema';
+export * from './callbackUrl.schema';
