+++ src/schemas/ingestion/index.ts
@@ -2,7 +2,7 @@ export * from './additionalParams.schema';
 export * from './inputFiles.schema';
 export * from './metadata.schema';
 export * from './polygonParts.schema';
-export * from './ingestionRequest.schema';
+export * from './layerData.schema';
 export * from './ingestionJobParams.schema';
 export * from './ingestionTaskParams.schema';
 export * from './layerNameFormats.schema';
