+++ src/schemas/ingestion/layerData.schema.ts
@@ -1,0 +1,6 @@
+import z from 'zod';
+import { inputFilesSchema } from './inputFiles.schema';
+
+export const layerDataSchema = z.object({
+  inputFiles: inputFilesSchema,
+});
