+++ src/types/ingestion/index.ts
@@ -1,1 +1,3 @@
-export * from './ingestion.type';
+export * from './layer.type';
+export * from './job.type';
+export * from './task.type';
