+++ src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './geo.utils';
 export * from './type.utils';
 export * from './layer.utils';
